Add files via upload
Good Old-Fashioned AI (GOFAI), also known as Symbolic AI, refers to a traditional approach to artificial intelligence that dominated the field in the early decades of its development. This approach is based on symbolic reasoning and manipulation of symbols to represent knowledge and solve problems.

Symbolic AI is an approach to artificial intelligence that emerged in the mid-20th century. It focuses on representing knowledge using symbols and employing formal logic for reasoning and decision-making. Unlike modern machine learning approaches that rely on statistical patterns, Symbolic AI emphasizes explicit programming of rules and knowledge structures.